+++ src/CommunityPatchLauncherFramework/Settings/Manager/SettingManager.cs
@@ -4,6 +4,7 @@ using CommunityPatchLauncherFramework.Settings.Writer;
 using System.Collections.Generic;
 using System.Linq;
 using System.IO;
+using System.Net.NetworkInformation;
 
 namespace CommunityPatchLauncherFramework.Settings.Manager
 {
@@ -113,11 +114,20 @@ namespace CommunityPatchLauncherFramework.Settings.Manager
         }
 
         /// <summary>
+        /// This will clear a value from the settings table
+        /// </summary>
+        /// <param name="key"></param>
+        public void ClearValue(string key)
+        {
+            settings?.RemoveWhere((pair) => pair.Key == key);
+        }
+
+        /// <summary>
         /// Reload all the settings
         /// </summary>
         public void Reload()
         {
-            settings = reader.LoadSettings(SettingFilePath);
+            settings = reader?.LoadSettings(SettingFilePath);
             settingsLoaded = true;
         }